+++ lib/whois/record/parser/whois.registrypro.pro.rb
@@ -7,64 +7,19 @@
 #++
 
 
-require 'whois/record/parser/base'
+require 'whois/record/parser/base_afilias'
 
 
 module Whois
   class Record
     class Parser
 
-      #
-      # = whois.registrypro.pro parser
-      #
       # Parser for the whois.registrypro.pro server.
+      # 
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
       #
-      # NOTE: This parser is just a stub and provides only a few basic methods
-      # to check for domain availability and get domain status.
-      # Please consider to contribute implementing missing methods.
-      # See WhoisNicIt parser for an explanation of all available methods
-      # and examples.
-      #
-      class WhoisRegistryproPro < Base
-
-        property_supported :status do
-          content_for_scanner.scan(/Status:(.*?)\n/).flatten
-        end
-
-        property_supported :available? do
-          !!(content_for_scanner =~ /No match/)
-        end
-
-        property_supported :registered? do
-          !available?
-        end
-
-
-        property_supported :created_on do
-          if content_for_scanner =~ /Created On:(.*)\n/
-            Time.parse($1)
-          end
-        end
-
-        property_supported :updated_on do
-          if content_for_scanner =~ /Last Updated On:(.*)\n/
-            Time.parse($1)
-          end
-        end
-
-        property_supported :expires_on do
-          if content_for_scanner =~ /Expiration Date:(.*)\n/
-            Time.parse($1)
-          end
-        end
-
-
-        property_supported :nameservers do
-          content_for_scanner.scan(/Name Server:(.+)\n/).flatten.map do |name|
-            Record::Nameserver.new(:name => name.downcase)
-          end
-        end
-
+      class WhoisRegistryproPro < BaseAfilias
       end
 
     end